Crate teth

Toy Ethereum implementation closely following the Yellow Paper (PDF). Structs, fields and methods are annotated with their formal definition where applicable.

Currently implements a small subset of the JSON-RPC API and a few custom methods. See the rpc module for details.

See README.md for more info.

Modules

rpc

Structs

AccountState
σ[a]

Block
B

BlockChain

BlockHeader
H

FEES
lazy_static for FeeSchedule

FeeSchedule
Fee Schedule (Appendix G)

ONE_ETHER
lazy_static for one Ether in Wei.

Options
CLI options (via structopt)

Signature
ECDSA signature

Transaction
T

WorldState
σ

Enums

SubCommand
CLI subcommands

Type Definitions

Wei
Alias for ethereum_types::U256