Struct teth::rpc::BlockResponse

pub struct BlockResponse {
    pub number: Option<u64>,
    pub hash: Option<H256>,
    pub parent_hash: H256,
    pub nonce: Option<u64>,
    pub logs_bloom: Option<Bloom>,
    pub transactions_root: H256,
    pub receipts_root: H256,
    pub miner: Address,
    pub difficulty: U256,
    pub total_difficulty: U256,
    pub extra_data: [u8; 32],
    pub size: usize,
    pub gas_limit: U256,
    pub gas_used: U256,
    pub timestamp: u64,
    pub full_transactions: Vec<TransactionResponse>,
    pub transaction_hashes: Vec<H256>,
    pub uncles: Vec<H256>,
}

Source: https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_getblockbyhash

Fields
number: Option<u64>

The block number. None when it's a pending block.

hash: Option<H256>parent_hash: H256nonce: Option<u64>

Hash of the generated proof-of-work. None when its pending block

logs_bloom: Option<Bloom>

The bloom filter for the logs of the block. None when its pending block.

transactions_root: H256

The root of the transaction trie of the block.

receipts_root: H256

The root of the receipts trie of the block.

miner: Address

The address of the beneficiary to whom the mining rewards were given.

difficulty: U256total_difficulty: U256extra_data: [u8; 32]size: usize

The size of this block in bytes

gas_limit: U256gas_used: U256timestamp: u64full_transactions: Vec<TransactionResponse>transaction_hashes: Vec<H256>uncles: Vec<H256>

Array of uncle hashes