Struct teth::rpc::TransactionResponse

pub struct TransactionResponse {
    pub block_hash: Option<H256>,
    pub number: Option<u64>,
    pub from: Address,
    pub gas: Wei,
    pub gas_price: Wei,
    pub hash: H256,
    pub input: Option<Vec<u8>>,
    pub nonce: U256,
    pub to: Option<Address>,
    pub transaction_index: Option<usize>,
    pub value: Wei,
    pub v: u8,
    pub r: U256,
    pub s: U256,
}

Source: https://github.com/ethereum/wiki/wiki/JSON-RPC#eth_gettransactionbyhash

Fields
block_hash: Option<H256>

Hash of the block where this transaction was in. None when it's pending.

number: Option<u64>

Block number where this transaction was in. None when it's pending.

from: Address

Address of the sender.

gas: Wei

Gas provided by the sender.

gas_price: Wei

Gas price provided by the sender in Wei

hash: H256

Hash of the transaction.

input: Option<Vec<u8>>

The data sent along with the transaction.

nonce: U256

The number of transactions made by the sender prior to this one.

to: Option<Address>

Address of the receiver. None when it's a contract creation transaction.

transaction_index: Option<usize>

Integer of the transaction's index position in the block. None when it's pending.

value: Wei

Value transferred in Wei.

v: u8

ECDSA recovery id

r: U256

ECDSA signature r

s: U256

ECDSA signature s

Methods

impl TransactionResponse

pub fn new(transaction: &Transaction, block: Option<&Block>) -> Self