Struct teth::Block

pub struct Block {
    pub header: BlockHeader,
    pub transactions: Vec<Transaction>,
    pub ommers: Vec<BlockHeader>,
}

B

The block in Ethereum is the collection of relevant pieces of information (known as the block header), H, together with information corresponding to the comprised transactions, T, and a set of other block headers U that are known to have a parent equal to the present block’s parent’s parent (such blocks are known as ommers).

Formally, we can refer to a block B:
B ≡ (BH, BT, BU)

Fields
header: BlockHeader

BH

transactions: Vec<Transaction>

BT

ommers: Vec<BlockHeader>

BU
Also known as 'uncles'.

Methods

impl Block

pub fn genesis_block() -> Self

Appendix I. Genesis Block

The genesis block is 15 items, and is specified thus:
((0256, KEC(RLP()), 0160 , stateRoot, 0, 0, 02048 , 217 , 0, 0, 3141592, time, 0, 0256 , KEC((42)), (), ())

Where 0256 refers to the parent hash, a 256-bit hash which is all zeroes; 0160 refers to the beneficiary address, a 160-bit hash which is all zeroes; 02048 refers to the log bloom, 2048-bit of all zeros; 217 refers to the difficulty; the transaction trie root, receipt trie root, gas used, block number and extradata are both 0, being equivalent to the empty byte array. The sequences of both ommers and transactions are empty and represented by (). KEC((42)) refers to the Keccak hash of a byte array of length one whose first and only byte is of value 42, used for the nonce. KEC(RLP()) value refers to the hash of the ommer list in RLP, both empty lists.

The proof-of-concept series include a development premine, making the state root hash some value stateRoot. Also time will be set to the initial timestamp of the genesis block. The latest documentation should be consulted for those values.

pub fn exodus_block() -> Self

Test block with transactions and genesis block as parent

pub fn to_rlp(&self) -> Vec<u8>

pub fn hash(&self) -> H256

Keccak 256-bit hash