Struct teth::Transaction

pub struct Transaction {
    pub nonce: U256,
    pub gas_price: Wei,
    pub gas_limit: U256,
    pub to: Option<Address>,
    pub value: Wei,
    pub signature: Signature,
    pub init: Option<Vec<u8>>,
    pub data: Option<Vec<u8>>,
}

T

A transaction (formally, T) is a single cryptographically-signed instruction constructed by an actor externally to the scope of Ethereum. While it is assumed that the ultimate external actor will be human in nature, software tools will be used in its construction and dissemination. There are two types of transactions: those which result in message calls and those which result in the creation of new accounts with associated code (known informally as ‘contract creation’).

Fields
nonce: U256

A scalar value equal to the number of transactions sent by the sender; formally Tn.

gas_price: Wei

A scalar value equal to the number of Wei to be paid per unit of gas for all computation costs incurred as a result of the execution of this transaction; formally Tp.

gas_limit: U256

A scalar value equal to the maximum amount of gas that should be used in executing this transaction. This is paid up-front, before any computation is done and may not be increased later; formally Tg.

to: Option<Address>

The 160-bit address of the message call’s recipient or, for a contract creation transaction, ∅, used here to denote the only member of B0; formally Tt.

value: Wei

A scalar value equal to the number of Wei to be transferred to the message call’s recipient or, in the case of contract creation, as an endowment to the newly created account; formally Tv.

signature: Signature

Values corresponding to the signature of the transaction and used to determine the sender of the transaction; formally Tw, Tr and Ts. This is expanded in Appendix F.

init: Option<Vec<u8>>

An unlimited size byte array specifying the EVM-code for the account initialisation procedure, formally Ti.
init is an EVM-code fragment; it returns the body, a second fragment of code that executes each time the account receives a message call (either through a transaction or due to the internal execution of code). init is executed only once at account creation and gets discarded immediately thereafter.
Empty for message calls.

data: Option<Vec<u8>>

An unlimited size byte array specifying the input data of the message call, formally Td.
Empty for contract creation transactions.

Methods

impl Transaction

pub fn sender(&self) -> Address

S

Appendix F specifies the function, S, which maps transactions to the sender, and happens through the ECDSA of the SECP-256k1 curve, using the hash of the transaction (excepting the latter three signature fields) as the datum to sign. For the present we simply assert that the sender of a given transaction T can be represented with S(T).

pub fn to_rlp(&self) -> Vec<u8>

pub fn hash(&self) -> H256

Keccak 256-bit hash

pub fn validate(&self)

Section 6 (beginning)

pub fn intrinsic_gas(&self) -> u64

Section 6.2: We define intrinsic gas g0, the amount of gas this transaction requires to be paid prior to execution, as follows:

pub fn up_front_cost(&self) -> U256

v0 (Equation 57)

pub fn execute(&self, state: &mut WorldState)

Section 6.2

pub fn execute_contract_creation(&self)

Section 7

pub fn execute_message_call(&self, state: &mut WorldState)

Section 8